platform/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
celery==5.4.0
redis==5.1.1
websocket-client==1.8.0
requests==2.32.3
boto3==1.35.29
python-dotenv==1.0.1
